+++ src/themes/catppuccin-macchiato.ts
@@ -1,0 +1,10 @@
+import { ThemeDef } from './types';
+
+const theme: ThemeDef = {
+    id: 'catppuccin-macchiato',
+    label: 'Catppuccin Macchiato',
+    group: 'Popular',
+    vars: { bg0Hard: '#181926', bg0: '#24273a', bg1: '#1e2030', bg2: '#363a4f', bg3: '#494d64', bg4: '#5b6078', fg0: '#cad3f5', fg1: '#b8c0e0', fg2: '#a5adcb', fg3: '#939ab7', fg4: '#8087a2', red: '#ed8796', green: '#a6da95', yellow: '#eed49f', blue: '#8aadf4', purple: '#c6a0f6', aqua: '#8bd5ca', orange: '#f5a97f' },
+};
+
+export default theme;