+++ src/themes/material-ocean.ts
@@ -1,0 +1,10 @@
+import { ThemeDef } from './types';
+
+const theme: ThemeDef = {
+    id: 'material-ocean',
+    label: 'Material Ocean',
+    group: 'Popular',
+    vars: { bg0Hard: '#0f111a', bg0: '#0f111a', bg1: '#090b10', bg2: '#1f2233', bg3: '#2e3250', bg4: '#444a73', fg0: '#eeffff', fg1: '#eeffff', fg2: '#c3cee3', fg3: '#8f93a2', fg4: '#464b5d', red: '#f07178', green: '#c3e88d', yellow: '#ffcb6b', blue: '#82aaff', purple: '#c792ea', aqua: '#89ddff', orange: '#f78c6c' },
+};
+
+export default theme;